Move sitemap definitions
