+++ services/graphql-server/src/graphql/definitions/sitemaps.js
@@ -1,0 +1,20 @@
+const gql = require('graphql-tag');
+
+module.exports = gql`
+
+type Mutation {
+  generateSitemaps(input: GenerateSitemapInput!): String!
+}
+
+enum SitemapType {
+  all
+  index
+  sections
+  content
+}
+
+input GenerateSitemapInput {
+  type: SitemapType!
+  baseUri: String!
+}
+`;
